dependency graph logic moved to manifest common
