+++ src/AppInstallerCLICore/Workflows/DependenciesFlow.h
@@ -50,133 +50,4 @@ namespace AppInstaller::CLI::Workflow
     // Inputs: PackageVersion, Manifest
     // Outputs: DependencySource
     void OpenDependencySource(Execution::Context& context);
-
-    struct DependencyGraph
-    {
-        DependencyGraph(AppInstaller::Manifest::Dependency root, 
-            AppInstaller::Manifest::DependencyList rootDependencies,
-            std::function<AppInstaller::Manifest::DependencyList(const AppInstaller::Manifest::Dependency&)> infoFunction) : m_root(root), getDependencies(infoFunction)
-        {
-            adjacents[m_root] = std::vector<AppInstaller::Manifest::Dependency>();
-            toCheck = std::vector<AppInstaller::Manifest::Dependency>();
-            rootDependencies.ApplyToType(AppInstaller::Manifest::DependencyType::Package, [&](AppInstaller::Manifest::Dependency dependency)
-                {
-                    toCheck.push_back(dependency);
-                    AddNode(dependency);
-                    AddAdjacent(root, dependency);
-                });
-        }
-
-        void BuildGraph()
-        {
-            if (toCheck.empty())
-            {
-                return;
-            }
-
-            for (int i = 0; i < toCheck.size(); ++i)
-            {
-                auto node = toCheck.at(i);
-
-                const auto& nodeDependencies = getDependencies(node); //TODO add error stream so we can report back
-
-                nodeDependencies.ApplyToType(AppInstaller::Manifest::DependencyType::Package, [&](AppInstaller::Manifest::Dependency dependency)
-                    {
-                        if (!HasNode(dependency))
-                        {
-                            toCheck.push_back(dependency);
-                            AddNode(dependency);
-                        }
-
-                        AddAdjacent(node, dependency);
-                    });
-            }
-
-            CheckForLoopsAndGetOrder();
-        }
-
-        void AddNode(AppInstaller::Manifest::Dependency node)
-        {
-            adjacents[node] = std::vector<AppInstaller::Manifest::Dependency>();
-
-        }
-
-        void AddAdjacent(AppInstaller::Manifest::Dependency node, AppInstaller::Manifest::Dependency adjacent)
-        {
-            adjacents[node].push_back(adjacent);
-        }
-
-        bool HasNode(AppInstaller::Manifest::Dependency dependency)
-        {
-            auto search = adjacents.find(dependency);
-            return search != adjacents.end();
-        }
-
-        bool HasLoop()
-        {
-            return hasLoop;
-        }
-        
-        void CheckForLoopsAndGetOrder()
-        {
-            installationOrder = std::vector<AppInstaller::Manifest::Dependency>();
-            std::set<AppInstaller::Manifest::Dependency> visited;
-            hasLoop = HasLoopDFS(visited, m_root);
-        }
-
-        std::vector<AppInstaller::Manifest::Dependency> GetInstallationOrder()
-        {
-            return installationOrder;
-        }
-
-        //-- only for debugging
-        void PrintOrder(AppInstaller::CLI::Execution::OutputStream info)
-        {
-            info << "order: ";
-            for (auto const& node : installationOrder)
-            {
-                info << node.Id << ", ";
-            }
-            info << std::endl;
-        }
-
-    private:
-        // TODO make HasLoopDFS iterative
-        bool HasLoopDFS(std::set<AppInstaller::Manifest::Dependency> visited, const AppInstaller::Manifest::Dependency& node)
-        {
-            visited.insert(node);
-            for (const auto& adjacent : adjacents.at(node))
-            {
-                auto search = visited.find(adjacent);
-                if (search == visited.end()) // if not found
-                {
-                    if (HasLoopDFS(visited, adjacent))
-                    {
-                        return true;
-                    }
-                }
-                else
-                {
-                    return true;
-                }
-            }
-
-            if (std::find(installationOrder.begin(), installationOrder.end(), node) == installationOrder.end())
-            {
-                installationOrder.push_back(node);
-            }
-
-            return false;
-        }
-
-        AppInstaller::Manifest::Dependency m_root;
-        std::map<AppInstaller::Manifest::Dependency, std::vector<AppInstaller::Manifest::Dependency>> adjacents; //(?) value should be a set instead of a vector?
-        std::function<AppInstaller::Manifest::DependencyList(const AppInstaller::Manifest::Dependency&)> getDependencies;
-        
-        bool hasLoop;
-        std::vector<AppInstaller::Manifest::Dependency> installationOrder;
-        
-        std::vector<AppInstaller::Manifest::Dependency> toCheck;
-        std::map<AppInstaller::Manifest::string_t, AppInstaller::Manifest::string_t> failedPackages;
-    };
 }
